// src/api/orval/model/emailBroadcastResponsesDtoEmailBroadcastTemplateDto.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface EmailBroadcastResponsesDtoEmailBroadcastTemplateDto {
  id?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  bodyHtml?: string | null;
  createDate?: string;
  lastUpdate?: string;
  /** @nullable */
  createdByUserId?: number | null;
}