// src/api/orval/model/emailBroadcastResponsesDtoEmailBroadcastTemplateSummaryDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface EmailBroadcastResponsesDtoEmailBroadcastTemplateSummaryDto {
  id?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  subject?: string | null;
  createDate?: string;
  lastUpdate?: string;
}