// src/api/orval/model/emailBroadcastResponsesDtoSentEmailLogDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface EmailBroadcastResponsesDtoSentEmailLogDto {
  id?: number;
  createDate?: string;
  /** @nullable */
  recipientUserId?: number | null;
  /** @nullable */
  recipientEmail?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  bodyHtml?: string | null;
  success?: boolean;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  sentByUserId?: number | null;
}