// src/api/orval/model/emailBroadcastResponsesGetEmailTemplatesResponse.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { EmailBroadcastResponsesDtoEmailBroadcastTemplateSummaryDto } from "./emailBroadcastResponsesDtoEmailBroadcastTemplateSummaryDto";

export interface EmailBroadcastResponsesGetEmailTemplatesResponse {
  /** @nullable */
  items?: EmailBroadcastResponsesDtoEmailBroadcastTemplateSummaryDto[] | null;
}