// src/api/orval/model/emailBroadcastResponsesGetSentEmailHistoryResponse.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { EmailBroadcastResponsesDtoSentEmailLogDto } from "./emailBroadcastResponsesDtoSentEmailLogDto";

export interface EmailBroadcastResponsesGetSentEmailHistoryResponse {
  page?: number;
  pageSize?: number;
  totalCount?: number;
  /** @nullable */
  items?: EmailBroadcastResponsesDtoSentEmailLogDto[] | null;
}